Bug 1154043 - (part 2) Build band limited tables after fundamental frequency is known r=dminor
If we build the band limited tables after the fundamental frequency is known,
we can exclude partials that are above the nyquist frequency for the sampling
frequency being used.

We rebuild the band limited tables each time we see a request for data for a
lower fundamental frequency so we have the required partials.

--HG--
extra : rebase_source : 757e5d3de6ed76cf45a3323e633292108e6b4cc0