Bug 1276110 - Workaround a wget bug by not performing internationalisation. r=ted
Taskclusters cloud-mirror sends a url-encoded response as a redirect, which allows it to point at the absolute url to the file in another AWS region and then it fetches and caches it while returning said file to our client (wget in this case)
Wget however is performing ASCII-->UTF8 conversion on the url on windows, while at the same time also performing a urldecodeon the redirected url.
This urldecode is breaking Taskclusters cloud-mirror, so we work around that by telling wget not to perform the conversion at all (we don't need charset conversion here)

Alternative option was to set --local-encoding=ASCII, however that option doesn't work on our current linux machines (and likely OSX builders, but untested there) while --no-iri works on all 3 platforms, so we are going with that.

MozReview-Commit-ID: EXmiDpozFNr

--HG--
extra : source : 37edfd2c53edf8829992fb418b0f27b676f5b830